requirements.txt
streamlit
langchain
chromadb
python-dotenv
sentence-transformers
pdfplumber
groq
